Move private plugin API to a separate file. (#52)
The libmptcpd library exposes an API that is meant solely for use by
the mptcpd program, not plugins.  Move functions in the private plugin
API to a new <mptcpd/plugin_private.h> that doesn't get installed.
