+++ include/mptcpd/Makefile.am
@@ -11,4 +11,5 @@ pkginclude_HEADERS =		\
 
 noinst_HEADERS =		\
 	config-private.h	\
-	path_manager_private.h
+	path_manager_private.h 	\
+	plugin_private.h
