+++ tests/test-plugin.c
@@ -19,7 +19,7 @@
 
 #include <ell/test.h>
 
-#include <mptcpd/plugin.h>
+#include <mptcpd/plugin_private.h>
 
 #include "test-plugin.h"
